Remove examples from native-component-list
