+++ apps/native-component-list/src/screens/Audio/AudioInputSelector.tsx
@@ -1,26 +1,25 @@
-import { Audio } from 'expo-av';
-import { RecordingInput } from 'expo-av/build/Audio/Recording.types';
+import { AudioRecorder, RecordingInput } from 'expo-audio';
 import React, { useCallback, useEffect, useState } from 'react';
 import { Text, View } from 'react-native';
 
-import ListButton from '../../../components/ListButton';
+import ListButton from '../../components/ListButton';
 
 type Props = {
-  recordingObject?: Audio.Recording;
+  recorder?: AudioRecorder;
 };
 
-function AudioInputSelector({ recordingObject }: Props) {
+function AudioInputSelector({ recorder }: Props) {
   const [availableInputs, setAvailableInputs] = useState<RecordingInput[]>([]);
   const [currentInput, setCurrentInput] = useState<RecordingInput | null>(null);
 
   const checkInputs = useCallback(async () => {
-    if (recordingObject) {
-      const availInputs = await recordingObject.getAvailableInputs();
+    if (recorder) {
+      const availInputs = recorder.getAvailableInputs();
       setAvailableInputs(availInputs);
-      const curtInput = await recordingObject.getCurrentInput();
+      const curtInput = await recorder.getCurrentInput();
       setCurrentInput(curtInput);
     }
-  }, [recordingObject]);
+  }, [recorder]);
 
   useEffect(() => {
     checkInputs();
@@ -38,7 +37,7 @@ function AudioInputSelector({ recordingObject }: Props) {
               key={`input-${input.uid}`}
               title={`${isSelected ? '✓ ' : ''}${title}`}
               onPress={async () => {
-                await recordingObject?.setInput(input.uid);
+                recorder?.setInput(input.uid);
                 checkInputs();
               }}
             />
