+++ apps/native-component-list/src/screens/Audio/Recorder.tsx
@@ -21,8 +21,8 @@ import {
 } from 'react-native';
 
 import AudioInputSelector from './AudioInputSelector';
-import Button from '../../../components/Button';
-import Colors from '../../../constants/Colors';
+import Button from '../../components/Button';
+import Colors from '../../constants/Colors';
 
 type RecorderProps = {
   onDone?: (uri: string) => void;
