--- apps/native-component-list/src/screens/Audio/expo-audio/AudioModeSelector.tsx
@@ -1,6 +1,0 @@
-// see https://github.com/Microsoft/TypeScript/issues/8328#issuecomment-219583152
-import { default as d } from './AudioModeSelector.ios';
-
-// eslint-disable-next-line
-export * from './AudioModeSelector.ios';
-export default d;
