--- apps/native-component-list/src/screens/Audio/expo-audio/AudioModeSelector.web.tsx
@@ -1,3 +1,0 @@
-export default function AudioModeSelector() {
-  return null;
-}
